+++ src/merenda/nimkit/view/viewbase.nim
@@ -130,16 +130,16 @@ type
     inputsDirty*: bool
 
   LayoutConstraint* = ref object
-    xFirstItem*: View
+    xFirstItemRef*: BackRef[View]
     xFirstAttribute*: LayoutAttribute
     xRelation*: LayoutRelation
-    xSecondItem*: View
+    xSecondItemRef*: BackRef[View]
     xSecondAttribute*: LayoutAttribute
     xMultiplier*: float32
     xConstant*: float32
     xPriority*: LayoutPriority
     xActive*: bool
-    xOwningView*: View
+    xOwningViewRef*: BackRef[View]
 
   LayoutTerm* = object
     item*: View